Event created

+++ db/migrate/20171223160614_create_products.rb
@@ -5,7 +5,7 @@ class CreateProducts < ActiveRecord::Migration[5.1]
       t.integer :price, default: 0
       t.string :description
       t.integer :quantity, default: 0
-      t.integer :rating, default: 0.0
+      t.float :rating, default: 0.0
       t.integer :rating_count, default: 0
       t.references :establishment, foreign_key: true
       t.references :combo, foreign_key: true
